# Conflicts:
#	ECNLogement/src/main/java/fr/centrale/nantes/ecnlogement/controllers/ApplicationTools.java
#	ECNLogement/src/main/java/fr/centrale/nantes/ecnlogement/controllers/LoginController.java

+++ ECNLogement/src/main/java/fr/centrale/nantes/ecnlogement/controllers/LogementController.java
@@ -5,7 +5,7 @@
  * Vianney de Ponthaud - Maxence Nicolet
  * ----------------------------------------- */
  
-//-*- coding: utf-8 -*-
+//-*- coding: utf-8 
 package fr.centrale.nantes.ecnlogement.controllers;
 
 import static fr.centrale.nantes.ecnlogement.controllers.ApplicationTools.getMethod;
@@ -46,6 +46,7 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Enumeration;
 import java.util.LinkedList;
@@ -100,7 +101,7 @@ public class LogementController {
             String targetDirectory = request.getServletContext().getRealPath("FichierRez");
             
             
-            if(fichierRez!=null){
+            if(fichierRez!=null&& fichierRez.length()>0){
                 try {
                     Path path = Paths.get(targetDirectory);
                     if (!Files.exists(path)) {
@@ -112,11 +113,34 @@ public class LogementController {
                     //Path destination = new File(targetDirectory).toPath();
                     String newFileName ="fichierRez.csv";
                     Path destination =path.resolve(newFileName);
-                    Files.copy(fichierRez.toPath(), destination);
-                    importCsvRez(fichierRez) ;
+                    //On regarde si le fichier e déjà été importé
+                    //S'il l'a été on supprime les données qui ont été importées
+                    //avec le premier import pour ne garder que les données du second
+                    if (Files.exists(destination)) { 
+                        Files.delete(destination);
+                        //FIXME supprimer les personnes associées aux élèves, truc bizarre lors d'un deuxième importph
+                        Collection<Eleve> elevesToDelete=eleveRepository.findByNumscei(-1);
+                        ArrayList<Personne> personnesToDelete=new ArrayList<>();
+                        for (Eleve eleve : elevesToDelete) {
+                            personnesToDelete.add(eleve.getPersonne());
+                        }
+                        eleveRepository.deleteAll(elevesToDelete);
+                        personneRepository.deleteAll(personnesToDelete);
+                        repository.deleteAll();
+                        Files.copy(fichierRez.toPath(), destination);
+                        importCsvRez(fichierRez) ;
+                    }else {
+                        Files.copy(fichierRez.toPath(), destination);
+                        importCsvRez(fichierRez) ;
+                    }
+                    
                 } catch (IOException ex) {
                     Logger.getLogger(AdminController.class.getName()).log(Level.SEVERE, null, ex);
                 }
+                finally{
+                    returned = handleLogementList(user);
+                }
+                
         }
         }   
         return returned;
@@ -379,9 +403,9 @@ public class LogementController {
                 // Get header
                 List<String> header = new LinkedList<>();
                 StringTokenizer st = new StringTokenizer(line, ";");
-                while ((st.hasMoreElements())&&header.size()<14) {
+                while ((st.hasMoreElements())&& (header.size()<14)) {
                     String name = st.nextToken().trim();
-                    header.add(ApplicationTools.removeAccentsAndSpecialCharacters(name.replaceAll("\\p{C}", "")));
+                    header.add(ApplicationTools.removeAccentsAndSpecialCharacters(name));
                 }
                 // Get lines
                 line = reader.readLine();
@@ -393,7 +417,7 @@ public class LogementController {
                     String elem = "";
                     while ((i < line.length())&&(lineValues.size()<14)) {
                         if (line.substring(i, i + 1).equals(";")) {
-                            lineValues.add(ApplicationTools.removeAccentsAndSpecialCharacters(elem.replaceAll("\\p{C}", "")));
+                            lineValues.add(ApplicationTools.removeAccentsAndSpecialCharacters(elem));
                             elem = "";
                         } else {
                             if (line.substring(i, i + 1).equals(",")) {
